refactor : make resolver package
