+++ src/main/java/core/mvc/resolver/ServletRequestArgumentResolver.java
@@ -1,18 +1,19 @@
-package core.mvc.tobe;
+package core.mvc.resolver;
 
 import core.mvc.MethodParameter;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ServletResponseArgumentResolver extends AbstractHandlerMethodArgumentResolver{
+public class ServletRequestArgumentResolver extends AbstractHandlerMethodArgumentResolver {
+
     @Override
     public boolean supports(MethodParameter parameter) {
-        return parameter.getType().equals(HttpServletResponse.class);
+        return parameter.getType().equals(HttpServletRequest.class);
     }
 
     @Override
     public Object getMethodArgument(MethodParameter parameter, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        return response;
+        return request;
     }
 }
