+++ src/main/java/core/mvc/resolver/ServletResponseArgumentResolver.java
@@ -1,19 +1,18 @@
-package core.mvc.tobe;
+package core.mvc.resolver;
 
 import core.mvc.MethodParameter;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ServletRequestArgumentResolver extends AbstractHandlerMethodArgumentResolver {
-
+public class ServletResponseArgumentResolver extends AbstractHandlerMethodArgumentResolver{
     @Override
     public boolean supports(MethodParameter parameter) {
-        return parameter.getType().equals(HttpServletRequest.class);
+        return parameter.getType().equals(HttpServletResponse.class);
     }
 
     @Override
     public Object getMethodArgument(MethodParameter parameter, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        return request;
+        return response;
     }
 }
