+++ src/test/java/core/mvc/tobe/PathVariableArgumentResolverTest.java
@@ -4,6 +4,7 @@ import core.annotation.web.PathVariable;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.mvc.MethodParameter;
+import core.mvc.resolver.PathVariableArgumentResolver;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
